Warning to not pull for handling the same_file merge conflict, formatting
During SEA 8 my students encountered a pitfall, because they learned about a pull before pushing to avoid merge conflicts which lead to a local resolve of the merge conflict.
I had to hard reset their progress to an earlier commit with them, so they could pick up with same_file again and not pull that time. I think adding the warning will prevent this issue in the future.